requirements.txt
Django-2.2.6
Pillow-6.2.1